requirements.txt
tkinter
ttkbootstrap>=1.0.0
pycryptodome>=3.15.0
